:art: reformat

+++ user/mmaptest.c
@@ -10,11 +10,9 @@ void mmap_test();
 void fork_test();
 char buf[BSIZE];
 
-#define MAP_FAILED ((char *) -1)
+#define MAP_FAILED ((char *)-1)
 
-int
-main(int argc, char *argv[])
-{
+int main(int argc, char *argv[]) {
   mmap_test();
   fork_test();
   printf("mmaptest: all tests succeeded\n");
@@ -23,9 +21,7 @@ main(int argc, char *argv[])
 
 char *testname = "???";
 
-void
-err(char *why)
-{
+void err(char *why) {
   printf("mmaptest: %s failed: %s, pid=%d\n", testname, why, getpid());
   exit(1);
 }
@@ -33,12 +29,10 @@ err(char *why)
 //
 // check the content of the two mapped pages.
 //
-void
-_v1(char *p)
-{
+void _v1(char *p) {
   int i;
-  for (i = 0; i < PGSIZE*2; i++) {
-    if (i < PGSIZE + (PGSIZE/2)) {
+  for (i = 0; i < PGSIZE * 2; i++) {
+    if (i < PGSIZE + (PGSIZE / 2)) {
       if (p[i] != 'A') {
         printf("mismatch at %d, wanted 'A', got 0x%x\n", i, p[i]);
         err("v1 mismatch (1)");
@@ -56,32 +50,25 @@ _v1(char *p)
 // create a file to be mapped, containing
 // 1.5 pages of 'A' and half a page of zeros.
 //
-void
-makefile(const char *f)
-{
+void makefile(const char *f) {
   int i;
-  int n = PGSIZE/BSIZE;
+  int n = PGSIZE / BSIZE;
 
   unlink(f);
   int fd = open(f, O_WRONLY | O_CREATE);
-  if (fd == -1)
-    err("open");
+  if (fd == -1) err("open");
   memset(buf, 'A', BSIZE);
   // write 1.5 page
-  for (i = 0; i < n + n/2; i++) {
-    if (write(fd, buf, BSIZE) != BSIZE)
-      err("write 0 makefile");
+  for (i = 0; i < n + n / 2; i++) {
+    if (write(fd, buf, BSIZE) != BSIZE) err("write 0 makefile");
   }
-  if (close(fd) == -1)
-    err("close");
+  if (close(fd) == -1) err("close");
 }
 
-void
-mmap_test(void)
-{
+void mmap_test(void) {
   int fd;
   int i;
-  const char * const f = "mmap.dur";
+  const char *const f = "mmap.dur";
   printf("mmap_test starting\n");
   testname = "mmap_test";
 
@@ -91,8 +78,7 @@ mmap_test(void)
   // file.
   //
   makefile(f);
-  if ((fd = open(f, O_RDONLY)) == -1)
-    err("open (1)");
+  if ((fd = open(f, O_RDONLY)) == -1) err("open (1)");
 
   printf("test mmap f\n");
   //
@@ -110,28 +96,22 @@ mmap_test(void)
   // of the file to be mapped. the last argument is the starting
   // offset in the file.
   //
-  char *p = mmap(0, PGSIZE*2, PROT_READ, MAP_PRIVATE, fd, 0);
-  if (p == MAP_FAILED)
-    err("mmap (1)");
+  char *p = mmap(0, PGSIZE * 2, PROT_READ, MAP_PRIVATE, fd, 0);
+  if (p == MAP_FAILED) err("mmap (1)");
   _v1(p);
-  if (munmap(p, PGSIZE*2) == -1)
-    err("munmap (1)");
+  if (munmap(p, PGSIZE * 2) == -1) err("munmap (1)");
 
   printf("test mmap f: OK\n");
 
   printf("test mmap private\n");
   // should be able to map file opened read-only with private writable
   // mapping
-  p = mmap(0, PGSIZE*2, PROT_READ | PROT_WRITE, MAP_PRIVATE, fd, 0);
-  if (p == MAP_FAILED)
-    err("mmap (2)");
-  if (close(fd) == -1)
-    err("close (1)");
+  p = mmap(0, PGSIZE * 2, PROT_READ | PROT_WRITE, MAP_PRIVATE, fd, 0);
+  if (p == MAP_FAILED) err("mmap (2)");
+  if (close(fd) == -1) err("close (1)");
   _v1(p);
-  for (i = 0; i < PGSIZE*2; i++)
-    p[i] = 'Z';
-  if (munmap(p, PGSIZE*2) == -1)
-    err("munmap (2)");
+  for (i = 0; i < PGSIZE * 2; i++) p[i] = 'Z';
+  if (munmap(p, PGSIZE * 2) == -1) err("munmap (2)");
 
   printf("test mmap private: OK\n");
 
@@ -139,13 +119,10 @@ mmap_test(void)
 
   // check that mmap doesn't allow read/write mapping of a
   // file opened read-only.
-  if ((fd = open(f, O_RDONLY)) == -1)
-    err("open (2)");
-  p = mmap(0, PGSIZE*3, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
-  if (p != MAP_FAILED)
-    err("mmap (3)");
-  if (close(fd) == -1)
-    err("close (2)");
+  if ((fd = open(f, O_RDONLY)) == -1) err("open (2)");
+  p = mmap(0, PGSIZE * 3, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
+  if (p != MAP_FAILED) err("mmap (3)");
+  if (close(fd) == -1) err("close (2)");
 
   printf("test mmap read-only: OK\n");
 
@@ -153,24 +130,19 @@ mmap_test(void)
 
   // check that mmap does allow read/write mapping of a
   // file opened read/write.
-  if ((fd = open(f, O_RDWR)) == -1)
-    err("open (3)");
-  p = mmap(0, PGSIZE*3, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
-  if (p == MAP_FAILED)
-    err("mmap (4)");
-  if (close(fd) == -1)
-    err("close (3)");
+  if ((fd = open(f, O_RDWR)) == -1) err("open (3)");
+  p = mmap(0, PGSIZE * 3, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
+  if (p == MAP_FAILED) err("mmap (4)");
+  if (close(fd) == -1) err("close (3)");
 
   // check that the mapping still works after close(fd).
   _v1(p);
 
   // write the mapped memory.
-  for (i = 0; i < PGSIZE*2; i++)
-    p[i] = 'Z';
+  for (i = 0; i < PGSIZE * 2; i++) p[i] = 'Z';
 
   // unmap just the first two of three pages of mapped memory.
-  if (munmap(p, PGSIZE*2) == -1)
-    err("munmap (3)");
+  if (munmap(p, PGSIZE * 2) == -1) err("munmap (3)");
 
   printf("test mmap read/write: OK\n");
 
@@ -178,25 +150,20 @@ mmap_test(void)
 
   // check that the writes to the mapped memory were
   // written to the file.
-  if ((fd = open(f, O_RDWR)) == -1)
-    err("open (4)");
-  for (i = 0; i < PGSIZE + (PGSIZE/2); i++){
+  if ((fd = open(f, O_RDWR)) == -1) err("open (4)");
+  for (i = 0; i < PGSIZE + (PGSIZE / 2); i++) {
     char b;
-    if (read(fd, &b, 1) != 1)
-      err("read (1)");
-    if (b != 'Z')
-      err("file does not contain modifications");
+    if (read(fd, &b, 1) != 1) err("read (1)");
+    if (b != 'Z') err("file does not contain modifications");
   }
-  if (close(fd) == -1)
-    err("close (4)");
+  if (close(fd) == -1) err("close (4)");
 
   printf("test mmap dirty: OK\n");
 
   printf("test not-mapped unmap\n");
 
   // unmap the rest of the mapped memory.
-  if (munmap(p+PGSIZE*2, PGSIZE) == -1)
-    err("munmap (4)");
+  if (munmap(p + PGSIZE * 2, PGSIZE) == -1) err("munmap (4)");
 
   printf("test not-mapped unmap: OK\n");
 
@@ -206,42 +173,27 @@ mmap_test(void)
   // mmap two files at the same time.
   //
   int fd1;
-  if((fd1 = open("mmap1", O_RDWR|O_CREATE)) < 0)
-    err("open (5)");
-  if(write(fd1, "12345", 5) != 5)
-    err("write (1)");
+  if ((fd1 = open("mmap1", O_RDWR | O_CREATE)) < 0) err("open (5)");
+  if (write(fd1, "12345", 5) != 5) err("write (1)");
   char *p1 = mmap(0, PGSIZE, PROT_READ, MAP_PRIVATE, fd1, 0);
-  if(p1 == MAP_FAILED)
-    err("mmap (5)");
-  if (close(fd1) == -1)
-    err("close (5)");
-  if (unlink("mmap1") == -1)
-    err("unlink (1)");
+  if (p1 == MAP_FAILED) err("mmap (5)");
+  if (close(fd1) == -1) err("close (5)");
+  if (unlink("mmap1") == -1) err("unlink (1)");
 
   int fd2;
-  if((fd2 = open("mmap2", O_RDWR|O_CREATE)) < 0)
-    err("open (6)");
-  if(write(fd2, "67890", 5) != 5)
-    err("write (2)");
+  if ((fd2 = open("mmap2", O_RDWR | O_CREATE)) < 0) err("open (6)");
+  if (write(fd2, "67890", 5) != 5) err("write (2)");
   char *p2 = mmap(0, PGSIZE, PROT_READ, MAP_PRIVATE, fd2, 0);
-  if(p2 == MAP_FAILED)
-    err("mmap (6)");
-  if (close(fd2) == -1)
-    err("close (6)");
-  if (unlink("mmap2") == -1)
-    err("unlink (2)");
+  if (p2 == MAP_FAILED) err("mmap (6)");
+  if (close(fd2) == -1) err("close (6)");
+  if (unlink("mmap2") == -1) err("unlink (2)");
 
-  if(memcmp(p1, "12345", 5) != 0)
-    err("mmap1 mismatch");
-  if(memcmp(p2, "67890", 5) != 0)
-    err("mmap2 mismatch");
+  if (memcmp(p1, "12345", 5) != 0) err("mmap1 mismatch");
+  if (memcmp(p2, "67890", 5) != 0) err("mmap2 mismatch");
 
-  if (munmap(p1, PGSIZE) == -1)
-    err("munmap (5)");
-  if(memcmp(p2, "67890", 5) != 0)
-    err("mmap2 mismatch (2)");
-  if (munmap(p2, PGSIZE) == -1)
-    err("munmap (6)");
+  if (munmap(p1, PGSIZE) == -1) err("munmap (5)");
+  if (memcmp(p2, "67890", 5) != 0) err("mmap2 mismatch (2)");
+  if (munmap(p2, PGSIZE) == -1) err("munmap (6)");
 
   printf("test mmap two files: OK\n");
 
@@ -252,46 +204,38 @@ mmap_test(void)
 // mmap a file, then fork.
 // check that the child sees the mapped file.
 //
-void
-fork_test(void)
-{
+void fork_test(void) {
   int fd;
   int pid;
-  const char * const f = "mmap.dur";
+  const char *const f = "mmap.dur";
 
   printf("fork_test starting\n");
   testname = "fork_test";
 
   // mmap the file twice.
   makefile(f);
-  if ((fd = open(f, O_RDONLY)) == -1)
-    err("open (7)");
-  if (unlink(f) == -1)
-    err("unlink (3)");
-  char *p1 = mmap(0, PGSIZE*2, PROT_READ, MAP_SHARED, fd, 0);
-  if (p1 == MAP_FAILED)
-    err("mmap (7)");
-  char *p2 = mmap(0, PGSIZE*2, PROT_READ, MAP_SHARED, fd, 0);
-  if (p2 == MAP_FAILED)
-    err("mmap (8)");
+  if ((fd = open(f, O_RDONLY)) == -1) err("open (7)");
+  if (unlink(f) == -1) err("unlink (3)");
+  char *p1 = mmap(0, PGSIZE * 2, PROT_READ, MAP_SHARED, fd, 0);
+  if (p1 == MAP_FAILED) err("mmap (7)");
+  char *p2 = mmap(0, PGSIZE * 2, PROT_READ, MAP_SHARED, fd, 0);
+  if (p2 == MAP_FAILED) err("mmap (8)");
 
   // read just 2nd page.
-  if(*(p1+PGSIZE) != 'A')
-    err("fork mismatch (1)");
+  if (*(p1 + PGSIZE) != 'A') err("fork mismatch (1)");
 
-  if((pid = fork()) < 0)
-    err("fork");
+  if ((pid = fork()) < 0) err("fork");
   if (pid == 0) {
     _v1(p1);
-    if (munmap(p1, PGSIZE) == -1) // just the first page
+    if (munmap(p1, PGSIZE) == -1)  // just the first page
       err("munmap (7)");
-    exit(0); // tell the parent that the mapping looks OK.
+    exit(0);  // tell the parent that the mapping looks OK.
   }
 
   int status = -1;
   wait(&status);
 
-  if(status != 0){
+  if (status != 0) {
     printf("fork_test failed\n");
     exit(1);
   }
